[NOW-589] Fixed unexpected blank screen after updating FW versions
This is because the some of node objects that needs to update are missing in the DeviceManager,
and then the method 'firstWhere' will throw an error.
(This may occur in the Pinnacle routers with CF production)

Solution: Save the node objects and their corresponding status into memory as a cache, which
can be used later if the node objects unexpectedly get lost during rebooting.